pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signdocs-brasil"
version = "1.1.0"
description = "Official Python SDK for the SignDocs Brasil API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "SignDocs Brasil", email = "dev@signdocs.com.br"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.28.0",
    "PyJWT>=2.6.0",
    "cryptography>=39.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "responses>=0.23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://signdocs.com.br"
Documentation = "https://docs.signdocs.com.br/sdk/python"
Repository = "https://github.com/signdocs/signdocs-brasil-python"

[tool.setuptools.packages.find]
include = ["signdocs_brasil*"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["signdocs_brasil"]

[tool.coverage.report]
fail_under = 80
